requirements.txt
requests==2.18.4
lxml==4.1.1
PyExecJS==1.5.0
